rename: Game에서 GameLobby로 변경
